GPU-2037 Prevent deletion of inherited content tags

+++ cms-core/src/main/java/com/gentics/contentnode/object/Content.java
@@ -129,7 +129,9 @@ public abstract class Content extends AbstractContentObject implements TagContai
 	@Override
 	public void delete(boolean force) throws NodeException {
 		for (ContentTag tag : getContentTags().values()) {
-			tag.delete();
+			if (!tag.isInherited()) {
+				tag.delete();
+			}
 		}
 		performDelete();
 	}
